requirements.txt
git+https://github.com/Christoph-Koschel/sirent/blob/master/sirent/__init__.py@rewrite
aiohttp==3.7.3
altgraph==0.17
async-timeout==3.0.1
attrs==20.3.0
discord.py=1.3.0
cffi==1.14.4
chardet==3.0.4
future==0.18.2
idna==3.1
multidict==5.1.0
pefile==2019.4.18
pycparser==2.20
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
PyNaCl==1.4.0
python-dotenv
pywin32-ctypes==0.2.0
six==1.15.0
typing-extensions==3.7.4.3
yarl==1.6.3
